load a specific bundle in a session config
I've extended the url scheme for session config generation
to support the specification of a specific bundle to load
instead of the most recent non-empty bundle.

This is only useful for OTrunk-style learner data.

* I've also fixed a bug in the refactorization of process logging.

An example of loading a specific bundle in a session config:

Offering 8918

  http://localhost:3001/13/offering/8918

The jnlp run link for Sam Fentress's workgroup:

  http://localhost:3001/13/offering/8918/jnlp/38085

Which results in this url for the standard SAIL/OTrunk session config:

  http://localhost:3001/13/workgroups/38085/bundles.xml

Which generates this in the PortfolioManagerService element:

  <object class="net.sf.sail.emf.launch.PortfolioManagerService">
    <void property="portfolioUrlProvider">
      <object class="net.sf.sail.emf.launch.XmlUrlStringProviderImpl">
        <void property="urlString">
          <string>http://localhost:3001/13/offering/8918/bundle/38085/1</string>
        </void>
      </object>
    </void>
    <void property="bundlePoster">
      <object class="net.sf.sail.emf.launch.BundlePoster">
        <void property="postUrl">
          <string>http://localhost:3001/13/offering/8918/bundle/38085/1</string>
        </void>
      </object>
    </void>
  </object>

To get the complete list of Bundles for Sam's workgroup:

  http://localhost:3001/13/workgroups/38085/bundles.xml

Sam is a developer at CC and has created 24 session bundles for this
offering -- one of which is Bundle id: 79736.

This url creates a session config that specifically loads Bundle 79736

  http://localhost:3001/13/offering/8918/config/38085/1/bundle/79736

Which generates this in the PortfolioManagerService element:

  <object class="net.sf.sail.emf.launch.PortfolioManagerService">
    <void property="portfolioUrlProvider">
      <object class="net.sf.sail.emf.launch.XmlUrlStringProviderImpl">
        <void property="urlString">
          <string>http://localhost:3001/13/bundle/79736.xml</string>
        </void>
      </object>
    </void>
  </object>

The current implementation does not allow posting bundles back when loading
a specific bundle so no bundlePoster element is generated in this session config.

git-svn-id: https://svn.concord.org/svn/sds/trunk@540 4e27850c-0488-45a1-9bd4-87dd9910adcb